Update mc
Added timer killer if user terminates Minecraft before time expires on it's own. i.e.; Previous to this addition if a user terminated Minecraft the terminal would still be running the 'mchammer' function until either 9pm or midnight as it was unaware that the game had been exited... Now the terminal prompt will be returned up to 10 seconds after the game is quit.