Limit kill to minecraft only
This way it will only kill processes with "minecraft" in their command line and ignore any other java programs (ex. Eclipse IDE, which may be used for a programming class).

Broken down:
 - ps prints the process id and command line for all processes
 - this gets piped to awk, which outputs the pids for any process (except awk itself) with the string "minecraft" anywhere in the command used to start it
 - kill then kills those processes

I tested this on Linux Mint 19.1 with the latest game and launcher versions.
The reason for formatting ps output with "-o pid,command" is in case of a username containing minecraft, otherwise ps would output the user and awk would match it, killing all processes owned by that user.